refactor(unit4): hoist the safetensors leaf helpers into the HF substrate
Eligibility, row projection, snapshot detection and repo-id recovery
were engine-neutral code living under the vLLM module. A second
safetensors engine now needs the same six functions, so they move to
discovery/hf_repos with the engine id as a parameter; the vLLM leaf
file goes away and its tests move with the code.
